Force Directed Graph Implemented - Needs Debugging/Troubleshooting
Currently the force directed graph is implemented and appears to be somewhat functioning. Nodes bias towards query terms that appear more frequently. The problem is that the clustering is too precise and there's no variability for each individual document. For example, given the query: USA Israel Egypt Japan, for our testing dataset, we can see that the nodes move towards USA, but there exist documents with a greater bias towards other terms than the USA - but these documents are moving in that direction anyway. Could be an unchecked force (we only want a link force) or some other bug?