Fixed Indexing
Added a novel dataset for testing (four corners). Debugged TextSearch.js so it now works as intended. Bounds check added to ClusterCloud. Both four corners and sampledata correctly plotting now as validation. Important to note that, as is, even though a lot of squares are colored blue the nodes are being correctly moved.