Minor Fixes, Basic Event Handling, Created Rough List of Remaining Tasks
Mouse events tentatively created. Cleaned up ClusterCloud.js code.

Checklist of things remaining to be completed (could be more, just a rough list):
1. animations (this might be something for Aaron to do)
2. on-click subset selection for zooming (check for 2 independently selected squares, stretch to scale)
3. higher/lower granularity (this is kind of already done - via a field that is set, but would need the code to generate the new layout once this value changes. Or an event trigger)
4. search term resizing depending on level of zoom (skew the word sizes to the perspective basically. IE zooming into the top left corner enlarges search terms near the top left, while those at the bottom right would be smaller by comparison [proportional to the distance away from the focal point of the window]).
5. pattern mapping instead of color mapping (free up color for other properties)
6. total number of documents displayed somewhere (can be above for the total, tooltips for individual sections)
7. Toggle to display the graph itself (documents depicted as circles placed between the query terms) or clustered variation.
8. control mechanism (IE 2 squares, a large one displaying the entire area and a smaller one that "highlights" a section of it that can be controlled and moved around to adjust the viewport on the cloud itself [the abstraction modifies the perspective on the cloud itself])

maybe:
1. enumerate all possible mappings for the search terms (layouts change depending on their ordering around the edge) and select the "optimal"(how?) for viewing.
2. automatic clustering for peaks (one idea: take a cutting plane at a certain %, draw contours along the cut sections)

Known bug with the implementation:
It appears that the lines composing the subsections of the cloud move upon a mouse event. Not really sure why, I only change their color. I was just going to abuse draw order to force the color of the line to be the appropriate color (the one in question would simply be drawn OVER the other one). There should be no linear transformation along that axis for the grid itself, so I don't know why it looks so convincing to me that it's off by a few pixels when I move my mouse around.

How to address it:
Depending on what's exactly happening, we can proceed accordingly. One solution is to draw the grid independendently, and just fill inbetween maintaining a mapping table between each edge segment and every space it helps to construct (recolor based off of mouseover on the regions between). It can be done, but it's a decent amount of work. Right now the lines just somewhat adjust by a pixel or two depending on the draw order of the squares. If I can figure out why the squares are actually (it looks like they are to me at least) translating along the x-axis for seemingly no reason I could probably fix it. I think (still do) that I don't need to do anything other than change draw order, still not fully sure why it appears to me that the squares move upon an event trigger. Honestly, it's good enough as is. But it is bothering me to leave something imperfect. Just not worth my time to fix this one is my opinion. I can get something working in spite of this, it'll just be ever so slightly off. I just wish to note it in case someone else does, I am aware of it but don't know another solution other than to redo some of the code to change the abstraction. Maybe after I address more pertinent tasks.